+++ github-api.php
@@ -2561,7 +2561,6 @@ function vipgoci_github_pr_review_submit(
 			'comments'	=> array(),
 		);
 
-
 		/*
 		 * For each issue reported, format
 		 * and prepare to be published on
@@ -2806,16 +2805,33 @@ function vipgoci_github_pr_review_submit(
 		}
 
 		/**
-		 * Format skipped files message if it validation has issues
+		 * Check if there're previous existent comments about the same files
 		 */
-		if ( ! empty( $results[ VIPGOCI_SKIPPED_FILES ][ $pr_number ]['issues'] ) ) {
+		$pr_reviews_commented = vipgoci_github_pr_reviews_get(
+			$repo_owner,
+			$repo_name,
+			$pr_number,
+			$github_token,
+			array(
+				'login' => 'myself',
+				'state' => array( 'COMMENTED', 'CHANGES_REQUESTED' )
+			)
+		);
+
+		$validation_message = get_validation_message_prefix( VIPGOCI_VALIDATION_MAXIMUM_LINES, $skip_large_files_limit );
+		$results[VIPGOCI_SKIPPED_FILES][ $pr_number ] = vipgo_skip_file_check_previous_pr_comments( $results[VIPGOCI_SKIPPED_FILES][ $pr_number ], $pr_reviews_commented, $validation_message );
+
+		/**
+		 * Format skipped files message if the validation has issues
+		 */
+		if ( 0 < $results[ VIPGOCI_SKIPPED_FILES ][ $pr_number ]['total'] ) {
 			vipgoci_markdown_comment_add_pagebreak(
 				$github_postfields['body']
 			);
 
 			$github_postfields[ 'body' ] .= vipgoci_get_skipped_files_message(
 				$results[ VIPGOCI_SKIPPED_FILES ][ $pr_number ],
-				$skip_large_files_limit
+				$validation_message
 			);
 		}