+++ tests/integration/Skeleton.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Vipgoci\Tests\Integration;
+
+require_once( __DIR__ . '/IncludesForTests.php' );
+
+use PHPUnit\Framework\TestCase;
+
+// phpcs:disable PSR1.Files.SideEffects
+
+final class Skeleton extends TestCase {
+	/**
+	 * @covers ::
+	 */
+	public function test (): void {
+	}
+}